+++ ft_printf/headers/parser.h
@@ -9,6 +9,6 @@
 
 char		*parse_flags(const char *s, t_flags *flags);
 char		*parse_pad(const char *s, t_flags *flags);
-char		*parse_type(const char *s, va_list ap, t_flags *flags, size_t *out_len);
+char		*parse_type(const char *s, va_list ap, t_flags *flags, int *out_len);
 
 #endif
